MNT: Move _sanitize_key helper from ora_remote into annex_utils
While so far this code has only been used within ORA, sanitization
of annex keys is universally useful. This commit moves it into a
more domain-agnostic part of the code base.
